+++ plugins/source/aws/resources/services/cloudwatchlogs/metric_filters.go
@@ -3,13 +3,14 @@ package cloudwatchlogs
 import (
 	"context"
 
+	"github.com/apache/arrow/go/v13/arrow"
 	"github.com/aws/aws-sdk-go-v2/aws"
 	"github.com/aws/aws-sdk-go-v2/aws/arn"
 	"github.com/aws/aws-sdk-go-v2/service/cloudwatchlogs"
 	"github.com/aws/aws-sdk-go-v2/service/cloudwatchlogs/types"
 	"github.com/cloudquery/cloudquery/plugins/source/aws/client"
-	"github.com/cloudquery/plugin-sdk/v2/schema"
-	"github.com/cloudquery/plugin-sdk/v2/transformers"
+	"github.com/cloudquery/plugin-sdk/v3/schema"
+	"github.com/cloudquery/plugin-sdk/v3/transformers"
 )
 
 func MetricFilters() *schema.Table {
@@ -24,12 +25,10 @@ func MetricFilters() *schema.Table {
 			client.DefaultAccountIDColumn(false),
 			client.DefaultRegionColumn(false),
 			{
-				Name:     "log_group_arn",
-				Type:     schema.TypeString,
-				Resolver: resolveMetricFilterLogGroupArn,
-				CreationOptions: schema.ColumnCreationOptions{
-					PrimaryKey: true,
-				},
+				Name:       "log_group_arn",
+				Type:       arrow.BinaryTypes.String,
+				Resolver:   resolveMetricFilterLogGroupArn,
+				PrimaryKey: true,
 			},
 		},
 	}