+++ plugins/source/aws/resources/services/ec2/ebs_volume_statuses.go
@@ -3,12 +3,13 @@ package ec2
 import (
 	"context"
 
+	"github.com/apache/arrow/go/v13/arrow"
 	"github.com/aws/aws-sdk-go-v2/aws"
 	"github.com/aws/aws-sdk-go-v2/service/ec2"
 	"github.com/aws/aws-sdk-go-v2/service/ec2/types"
 	"github.com/cloudquery/cloudquery/plugins/source/aws/client"
-	"github.com/cloudquery/plugin-sdk/v2/schema"
-	"github.com/cloudquery/plugin-sdk/v2/transformers"
+	"github.com/cloudquery/plugin-sdk/v3/schema"
+	"github.com/cloudquery/plugin-sdk/v3/transformers"
 )
 
 func EbsVolumesStatuses() *schema.Table {
@@ -23,12 +24,10 @@ func EbsVolumesStatuses() *schema.Table {
 			client.DefaultAccountIDColumn(false),
 			client.DefaultRegionColumn(false),
 			{
-				Name:     "volume_arn",
-				Type:     schema.TypeString,
-				Resolver: resolveEbsVolumeStatusArn,
-				CreationOptions: schema.ColumnCreationOptions{
-					PrimaryKey: true,
-				},
+				Name:       "volume_arn",
+				Type:       arrow.BinaryTypes.String,
+				Resolver:   resolveEbsVolumeStatusArn,
+				PrimaryKey: true,
 			},
 		},
 	}