+++ plugins/source/aws/resources/services/lightsail/buckets.go
@@ -3,12 +3,15 @@ package lightsail
 import (
 	"context"
 
+	"github.com/apache/arrow/go/v13/arrow"
+	sdkTypes "github.com/cloudquery/plugin-sdk/v3/types"
+
 	"github.com/aws/aws-sdk-go-v2/aws"
 	"github.com/aws/aws-sdk-go-v2/service/lightsail"
 	"github.com/aws/aws-sdk-go-v2/service/lightsail/types"
 	"github.com/cloudquery/cloudquery/plugins/source/aws/client"
-	"github.com/cloudquery/plugin-sdk/v2/schema"
-	"github.com/cloudquery/plugin-sdk/v2/transformers"
+	"github.com/cloudquery/plugin-sdk/v3/schema"
+	"github.com/cloudquery/plugin-sdk/v3/transformers"
 )
 
 func Buckets() *schema.Table {
@@ -24,12 +27,12 @@ func Buckets() *schema.Table {
 			client.DefaultRegionColumn(false),
 			{
 				Name:     "able_to_update_bundle",
-				Type:     schema.TypeBool,
+				Type:     arrow.FixedWidthTypes.Boolean,
 				Resolver: schema.PathResolver("AbleToUpdateBundle"),
 			},
 			{
 				Name:     "tags",
-				Type:     schema.TypeJSON,
+				Type:     sdkTypes.ExtensionTypes.JSON,
 				Resolver: client.ResolveTags,
 			},
 		},