+++ plugins/source/aws/resources/services/rds/engine_versions.go
@@ -3,11 +3,14 @@ package rds
 import (
 	"context"
 
+	sdkTypes "github.com/cloudquery/plugin-sdk/v3/types"
+
+	"github.com/apache/arrow/go/v13/arrow"
 	"github.com/aws/aws-sdk-go-v2/service/rds"
 	"github.com/aws/aws-sdk-go-v2/service/rds/types"
 	"github.com/cloudquery/cloudquery/plugins/source/aws/client"
-	"github.com/cloudquery/plugin-sdk/v2/schema"
-	"github.com/cloudquery/plugin-sdk/v2/transformers"
+	"github.com/cloudquery/plugin-sdk/v3/schema"
+	"github.com/cloudquery/plugin-sdk/v3/transformers"
 )
 
 func EngineVersions() *schema.Table {
@@ -22,16 +25,14 @@ func EngineVersions() *schema.Table {
 			client.DefaultAccountIDColumn(true),
 			client.DefaultRegionColumn(true),
 			{
-				Name:     "_engine_version_hash",
-				Type:     schema.TypeString,
-				Resolver: client.ResolveObjectHash,
-				CreationOptions: schema.ColumnCreationOptions{
-					PrimaryKey: true,
-				},
+				Name:       "_engine_version_hash",
+				Type:       arrow.BinaryTypes.String,
+				Resolver:   client.ResolveObjectHash,
+				PrimaryKey: true,
 			},
 			{
 				Name:     "tag_list",
-				Type:     schema.TypeJSON,
+				Type:     sdkTypes.ExtensionTypes.JSON,
 				Resolver: client.ResolveTagField("TagList"),
 			},
 		},