+++ plugins/source/aws/resources/services/s3/bucket_encryption_rules.go
@@ -3,12 +3,14 @@ package s3
 import (
 	"context"
 
+	"github.com/apache/arrow/go/v13/arrow"
 	"github.com/aws/aws-sdk-go-v2/service/s3"
 	"github.com/aws/aws-sdk-go-v2/service/s3/types"
 	"github.com/cloudquery/cloudquery/plugins/source/aws/client"
 	"github.com/cloudquery/cloudquery/plugins/source/aws/resources/services/s3/models"
-	"github.com/cloudquery/plugin-sdk/v2/schema"
-	"github.com/cloudquery/plugin-sdk/v2/transformers"
+	"github.com/cloudquery/plugin-sdk/v3/scalar"
+	"github.com/cloudquery/plugin-sdk/v3/schema"
+	"github.com/cloudquery/plugin-sdk/v3/transformers"
 )
 
 func bucketEncryptionRules() *schema.Table {
@@ -21,7 +23,7 @@ func bucketEncryptionRules() *schema.Table {
 			client.DefaultAccountIDColumn(false),
 			{
 				Name:     "bucket_arn",
-				Type:     schema.TypeString,
+				Type:     arrow.BinaryTypes.String,
 				Resolver: schema.ParentColumnResolver("arn"),
 			},
 		},
@@ -32,12 +34,12 @@ func fetchS3BucketEncryptionRules(ctx context.Context, meta schema.ClientMeta, p
 	r := parent.Item.(*models.WrappedBucket)
 	cl := meta.(*client.Client)
 	svc := cl.Services().S3
-	region := parent.Get("region").(*schema.Text)
+	region := parent.Get("region").(*scalar.String)
 	if region == nil {
 		return nil
 	}
 	aclOutput, err := svc.GetBucketEncryption(ctx, &s3.GetBucketEncryptionInput{Bucket: r.Name}, func(o *s3.Options) {
-		o.Region = region.Str
+		o.Region = region.Value
 	})
 	if err != nil {
 		if client.IsAWSError(err, "ServerSideEncryptionConfigurationNotFoundError") {